requirements.txt
arxiv==2.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
feedparser==6.0.10
filelock==3.14.0
idna==3.7
numpy==1.26.4
pandas==2.2.2
platformdirs==4.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
tzdata==2024.1
uritools==4.0.2
urlextract==1.9.0
urllib3==2.2.1
